fix: introduce visualization switcher highlighting
- create common highlighting hook
- also add support for rich text widget

JIRA: F1-997
risk: low
