+++ libs/sdk-ui-dashboard/src/model/types/commonTypes.ts
@@ -8,6 +8,7 @@ import {
     ISettings,
     ISeparators,
     IEntitlementDescriptor,
+    Identifier,
 } from "@gooddata/sdk-model";
 import { ILocale } from "@gooddata/sdk-ui";
 import keys from "lodash/keys.js";
@@ -253,12 +254,14 @@ export interface DashboardFocusObject {
      * This means, that target widget will be focused in the dashboard,
      * and filter context will be set to the one used in the automation.
      */
-    automationId?: string;
+    automationId?: Identifier;
 
     /**
      * @beta
      * If provided, the dashboard will be opened in the context of the given widget.
      * This means, that target widget will be focused in the dashboard.
+     *
+     * @remarks This property is not a formal MD identifier
      */
     widgetId?: string;
 
@@ -267,7 +270,7 @@ export interface DashboardFocusObject {
      * If provided, the dashboard will be opened in the context of the given visualization.
      * This means, that target visualizations represented by widgets will be focused in the dashboard.
      */
-    visualizationId?: string;
+    visualizationId?: Identifier;
 }
 
 /**
