+++ libs/sdk-ui-dashboard/src/presentation/widget/widget/RichTextWidget/DefaultDashboardRichTextWidget.tsx
@@ -1,9 +1,10 @@
-// (C) 2020-2024 GoodData Corporation
+// (C) 2020-2025 GoodData Corporation
 import React from "react";
 import cx from "classnames";
 import { DashboardItem, DashboardItemBase } from "../../../presentationComponents/index.js";
 import { IDefaultDashboardRichTextWidgetProps } from "./types.js";
 import { DashboardRichText } from "../../richText/DashboardRichText.js";
+import { useWidgetHighlighting } from "../../common/useWidgetHighlighting.js";
 
 /**
  * @internal
@@ -13,10 +14,15 @@ export const DefaultDashboardRichTextWidget: React.FC<IDefaultDashboardRichTextW
     screen,
     dashboardItemClasses,
 }) => {
+    const { elementRef, highlighted } = useWidgetHighlighting(widget);
+
     return (
         <DashboardItem
-            className={cx(dashboardItemClasses, "type-visualization", "gd-dashboard-view-widget")}
+            className={cx(dashboardItemClasses, "type-visualization", "gd-dashboard-view-widget", {
+                "gd-highlighted": highlighted,
+            })}
             screen={screen}
+            ref={elementRef}
         >
             <DashboardItemBase visualizationClassName="gd-rich-text-widget-wrapper">
                 {() => <DashboardRichText widget={widget} />}
