+++ libs/sdk-ui-dashboard/src/presentation/widget/widget/VisualizationSwitcherWidget/DefaultDashboardVisualizationSwitcherWidget.tsx
@@ -1,4 +1,4 @@
-// (C) 2020-2024 GoodData Corporation
+// (C) 2020-2025 GoodData Corporation
 
 import React from "react";
 import cx from "classnames";
@@ -6,6 +6,7 @@ import cx from "classnames";
 import { DashboardItem, DashboardItemBase } from "../../../presentationComponents/index.js";
 import { IDefaultDashboardVisualizationSwitcherWidgetProps } from "./types.js";
 import { DashboardVisualizationSwitcher } from "../../visualizationSwitcher/DashboardVisualizationSwitcher.js";
+import { useWidgetHighlighting } from "../../common/useWidgetHighlighting.js";
 
 /**
  * @internal
@@ -13,10 +14,15 @@ import { DashboardVisualizationSwitcher } from "../../visualizationSwitcher/Dash
 export const DefaultDashboardVisualizationSwitcherWidget: React.FC<
     IDefaultDashboardVisualizationSwitcherWidgetProps
 > = ({ widget, screen, dashboardItemClasses }) => {
+    const { elementRef, highlighted } = useWidgetHighlighting(widget);
+
     return (
         <DashboardItem
-            className={cx(dashboardItemClasses, "type-visualization", "gd-dashboard-view-widget")}
+            className={cx(dashboardItemClasses, "type-visualization", "gd-dashboard-view-widget", {
+                "gd-highlighted": highlighted,
+            })}
             screen={screen}
+            ref={elementRef}
         >
             <DashboardItemBase visualizationClassName="gd-visualization-switcher-widget-wrapper">
                 {() => <DashboardVisualizationSwitcher widget={widget} screen={screen} />}
